rewrite plotly-factory to TS vue component
