+++ src/helpers/shadow-copy.ts
@@ -1,5 +1,6 @@
 import { deserialize, serialize } from '@/helpers/units/parseObject';
-import { get, set } from 'lodash';
+import get from 'lodash/get';
+import set from 'lodash/set';
 
 /**
  * Use ShadowMapping to bind sources of values to inputs used in the component.
